standalone: no-INT64 shipping path + CLI robustness + release harness
- no-INT64 (shipping INT path now free of native 64-bit types,
  `make check-no-int64` PASS):
  * galosh_raw_cpu_int.c / galosh_int_p0.cl: run_max*3/5 threshold via int32
    split-multiply -- exact for run_max>=0, verified BIT-EXACT pre/post on
    real SIDD frames (r32 CPU maxdiff 0; i16 GPU output ndiff 0)
  * galosh_int_p0/p1/p2.cl: flattened loop indices long->int (fit int32 for
    any real frame; commented)
  * profiling counters / opcount pixel-arithmetic marked `no64-exempt`
- galosh_gpu.h: kernel/source loader now falls back to <exe dir>/<basename>,
  so every GPU CLI works from ANY working directory (was CWD-only after the
  path anonymization; verified from C:/tmp)
- galosh_raw_gpu.c: honor externally supplied (alpha, sigma^2) -- the CLI
  accepted them but Phase-0 readback silently overwrote them; now positive
  values override the model slots on both the o32 and legacy paths (CPU-exe
  parity, verified 82.7 dB agreement vs CPU with the same external model)
- release harness: Makefile + portable build.sh (raw/yuv/gpu/test/bench-small/
  check-no-int64), check_no_int64.sh (scoped scanner with documented
  exemptions), tests/run_smoke.sh (19 cases: constant/random/odd/small/
  high-noise/near-black x RAW FP32+INT32 / YUV; GPU smoke with device probe),
  tests/run_bench_small.sh (seeded synthetic PG pair, dataset-free, CSV/JSON)
  -- all PASS

+++ standalone/galosh_int_p1.cl
@@ -49,9 +49,10 @@ void k_p1_sigma_ch(__global const lbuf_t *gat_q20, int width, int height,
   /* horizontal Laplacian: y in [0,halfheight), x in [0,halfwidth-2) */
   const int nxh = halfwidth - 2;
   if(nxh > 0) {
-    long tot = (long)halfheight * nxh;
-    for(long t = tid; t < tot; t += P1_WG) {
-      int y = (int)(t / nxh), x = (int)(t - (long)y * nxh);
+    /* no-INT64: hh*nxh <= W*H/4 < 2^31 for any real frame */
+    int tot = halfheight * nxh;
+    for(int t = tid; t < tot; t += P1_WG) {
+      int y = t / nxh, x = t - y * nxh;
       int r = 2*y + dy0, c0 = 2*x + dx0, c1 = 2*(x+1) + dx0, c2 = 2*(x+2) + dx0;
       if(r >= height || c2 >= width) continue;
       int v0 = LDB(gat_q20, (size_t)r * width + c0);
@@ -67,9 +68,10 @@ void k_p1_sigma_ch(__global const lbuf_t *gat_q20, int width, int height,
   /* vertical Laplacian: y in [0,halfheight-2), x in [0,halfwidth) */
   const int nyv = halfheight - 2;
   if(nyv > 0) {
-    long tot = (long)nyv * halfwidth;
-    for(long t = tid; t < tot; t += P1_WG) {
-      int y = (int)(t / halfwidth), x = (int)(t - (long)y * halfwidth);
+    /* no-INT64: nyv*hw <= W*H/4 < 2^31 for any real frame */
+    int tot = nyv * halfwidth;
+    for(int t = tid; t < tot; t += P1_WG) {
+      int y = t / halfwidth, x = t - y * halfwidth;
       int c = 2*x + dx0, r0 = 2*y + dy0, r1 = 2*(y+1) + dy0, r2 = 2*(y+2) + dy0;
       if(r2 >= height || c >= width) continue;
       int v0 = LDB(gat_q20, (size_t)r0 * width + c);
